requirements.txt
fastapi
uvicorn
numpy
opencv-python
insightface
onnxruntime
requests
fastapi
uvicorn
python-dotenv
cloudinaryqrcode[pil]
pillow